Modified app.css
In the previous version , when switching to fullscreen mode ,
the search box would show however.

This is caused by the navbar taking space in non-fullscreen ,
but hidden in full screen mode, cause the css to be buggy.

Modified transition of .electronSearchText-box to all , add opacity 0,
and add opacity 1 to .electronSearchText-visible

You can still click on the search box accidently on the top right,
but this is the way I thought to modify the least code.